rpi/requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
matplotlib==3.8.2
python-dotenv==1.0.0 